Added unified drawcall support. (#10)
This adds a unified draw call to the BGFX API that wraps a number of C functions associated with performing a rendering draw call. This vastly improves the performance of the rendering loop.

This also refactors the BGFX handle representation in node to be a uint32_t, and makes them fully unmanages, which is considerably easier from a bookkeeping standpoint. Handles are also now annotated by type to allow better error handling.